Patch test tri3 fem passes

+++ veamy/include/veamy/models/constraints/structures/Angle.h
@@ -8,7 +8,7 @@ struct Angle{
     double angle;
 
     Angle(double a){
-        a = a - 360.0*(a/360);
+        a = a - 360.0*int(a/360);
         if(a>=0 && a<180) {
             angle = a;
         }else{
